lang-graph/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
openai>=1.50.0
anthropic>=0.40.0
